Added Character classes and fixed project bugs that kept it from running - Fixes #13 - Fixes #21
