+++ Model/Entities/Equipment.cs
@@ -1,0 +1,20 @@
+using Model.Entities;
+
+namespace Model.Entities
+{
+	public enum EquipmentType
+	{
+		HEAD = 0,
+
+		HAND,
+
+		SHOE
+	}
+
+	public class Equipment
+	{
+		public Stats EquipmentStats { get; }
+
+		public EquipmentType EquipmentType { get; }
+	}
+}
