+++ Model/Entities/Type.cs
@@ -1,0 +1,39 @@
+namespace Model.Entities
+{
+	public enum Type
+	{
+		NORMAL = 0,
+
+		FIRE,
+
+		WATER,
+
+		GRASS,
+
+		ELECTRIC,
+
+		ICE,
+
+		FIGHTING,
+
+		POISON,
+
+		GROUND,
+
+		FLYING,
+
+		PSYCHIC,
+
+		BUG,
+
+		ROCK,
+
+		GHOST,
+
+		DRAGON,
+
+		DARK,
+
+		STEEL
+	}
+}
