--- Model/Entity/Equipment.cs
@@ -1,20 +1,0 @@
-using Model.Entity.Stats;
-
-namespace Model.Entity.Equipment
-{
-    public enum EquipmentType
-    {
-        HEAD = 0,
-
-        HAND,
-
-        SHOE
-    }
-
-    public class Equipment
-    {
-        public Stats EquipmentStats { get; }
-
-        public EquipmentType EquipmentType { get; }
-    }
-}
