--- Model/Entity/Stats.cs
@@ -1,17 +1,0 @@
-namespace Model.Entity.Stats
-{
-    public class Stats
-    {
-        public int Attack { get; set; }
-
-        public int SpecialAttack { get; set; }
-
-        public int Defense { get; set; }
-
-        public int SpecialDefense { get; set; }
-
-        public int Speed { get; set; }
-
-        public int MaxHP { get; set; }
-    }
-}
