--- Model/Entity/Type.cs
@@ -1,39 +1,0 @@
-namespace Model.Entity.Type
-{
-    public enum Type
-    {
-        NORMAL = 0,
-
-        FIRE,
-
-        WATER,
-
-        GRASS,
-
-        ELECTRIC,
-
-        ICE,
-
-        FIGHTING,
-
-        POISON,
-
-        GROUND,
-
-        FLYING,
-
-        PSYCHIC,
-
-        BUG,
-
-        ROCK,
-
-        GHOST,
-
-        DRAGON,
-
-        DARK,
-
-        STEEL
-    }
-}
