+++ Model/Game/GameStates.cs
@@ -1,26 +1,26 @@
 namespace Model.Game
 {
-    public static enum GameStates
-    {
-        TITLE_SCREEN = 0,
+	public enum GameStates
+	{
+		TITLE_SCREEN = 0,
 
-        //Gameplay 
-        FREE_ROAM,
+		//Gameplay 
+		FREE_ROAM,
 
-        WILD_POKEMON_BATTLE,
+		WILD_POKEMON_BATTLE,
 
-        TRAINER_POKEMON_BATTLE,
+		TRAINER_POKEMON_BATTLE,
 
-        PAUSE,
+		PAUSE,
 
-        // Menus
-        POKEMON_MENU,
+		// Menus
+		POKEMON_MENU,
 
-        INVENTORY_MENU,
+		INVENTORY_MENU,
 
-        // Fade to black when loading an area, then fade in to the scene
-        FADE_OUT,
+		// Fade to black when loading an area, then fade in to the scene
+		FADE_OUT,
 
-        FADE_IN
-    }
+		FADE_IN
+	}
 }
